requirements.txt
selenium
webdriver-manager
playwright
python-dotenv
